procedural generation triggers endless dig moved player class to tank
